Improve NobtLoader
We move all the logic of rendering and tracking the fetchStatus of the
nobt to the NobtLoader itself. This allows us to generically display the
progress bar and switch to the underlying component as soon as the nobt
finished loading.
