+++ src/components/NobtLoader/NobtLoader.scss
@@ -1,0 +1,16 @@
+.loader {
+  width: 100%;
+  height: 80vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+
+  .separator {
+    flex-grow: 1;
+  }
+
+  .progressBar {
+    flex-grow: 3;
+  }
+}
