+++ src/components/NobtLoader/index.js
@@ -1,0 +1,3 @@
+import withNobtLoader from "./withNobtLoader";
+
+export default withNobtLoader
