+++ src/routes/App/components/Nobt.scss
@@ -56,20 +56,3 @@
     }
   }
 }
-
-.loader {
-  width: 100%;
-  height: 80vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-
-  .separator {
-    flex-grow: 1;
-  }
-
-  .progressBar {
-    flex-grow: 3;
-  }
-}
