+++ src/routes/App/containers/NobtContainer.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 import Nobt from "../components/Nobt";
 import { addMember, invalidateNobt } from "../modules/currentNobt/actions";
-import { getName, getCurrency, getMembers, getBills, getTotal, getFilteredBills, getFetchNobtStatus, getCreatedOn, isNobtEmpty } from "../modules/currentNobt/selectors";
+import { getCreatedOn, getCurrency, getFilteredBills, getMembers, getName, getTotal, isNobtEmpty } from "../modules/currentNobt/selectors";
 import { updateBillFilter, updateBillSortProperty } from "../modules/viewState/actions";
 import { getBillFilter, getBillSortProperty } from "../modules/viewState/selectors";
 import withNavigation from "../../../components/hoc/withNavigation";
@@ -15,7 +15,6 @@ const mapStateToProps = (state) => {
     bills: getFilteredBills(state),
     billFilter: getBillFilter(state),
     billSortProperty: getBillSortProperty(state),
-    fetchStatus: getFetchNobtStatus(state),
     createdOn: getCreatedOn(state),
     isNobtEmpty: isNobtEmpty(state)
   };
@@ -27,7 +26,7 @@ const mapDispatchToProps = (dispatch, props) => {
     updateBillFilter: (memberName) => dispatch(updateBillFilter(memberName)),
     updateBillSortProperty: (property) => dispatch(updateBillSortProperty(property)),
     invalidateNobtData: () => dispatch(invalidateNobt())
-  }
+  };
 };
 
-export default withNavigation(connect(mapStateToProps, mapDispatchToProps)(Nobt))
+export default withNavigation(connect(mapStateToProps, mapDispatchToProps)(Nobt));
